+++ src/app/shared/core/migrating-dialog/migrating-dialog.component.ts
@@ -1,0 +1,8 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-migrating-dialog',
+  templateUrl: './migrating-dialog.component.html',
+  styleUrls: ['./migrating-dialog.component.scss'],
+})
+export class MigratingDialogComponent {}