+++ src/app/shared/services/database/table/capacitor-filesystem-table/capacitor-filesystem-table.ts
@@ -142,16 +142,16 @@ export class CapacitorFilesystemTable<T extends Tuple> implements Table<T> {
     });
   }
 
-  async update(newTuple: T, comparator: (x: T, y: T) => boolean) {
+  async update(tuples: T[], comparator: (x: T, y: T) => boolean) {
     return this.mutex.runExclusive(async () => {
       const afterDeletion = differenceWith(
         this.tuples$.value,
-        [newTuple],
+        tuples,
         comparator
       );
-      this.tuples$.next(afterDeletion.concat(newTuple));
+      this.tuples$.next(afterDeletion.concat(tuples));
       await this.dumpJson();
-      return newTuple;
+      return tuples;
     });
   }