+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.15.4 - 2021-01-26
+
+### Fixed
+
+- Migrate `Proof.diaBackendAssetId`. #482
+- Determine if pre-fetch by remote asset count. #483
+
+### Changed
+
+#### UI/UX
+
+- Change tab labels to icons. #485
+
 ## 0.15.3 - 2021-01-25
 
 ### Fixed